+++ src/main/java/wooteco/subway/admin/controller/LineStationController.java
@@ -5,15 +5,19 @@ import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-import wooteco.subway.admin.dto.LineDetailResponse;
-import wooteco.subway.admin.dto.LineStationCreateRequest;
-import wooteco.subway.admin.service.LineService;
+import wooteco.subway.admin.dto.request.LineStationCreateRequest;
+import wooteco.subway.admin.dto.resopnse.ApiError;
+import wooteco.subway.admin.dto.resopnse.LineDetailResponse;
+import wooteco.subway.admin.exception.LineStationException;
+import wooteco.subway.admin.exception.NotFoundException;
 import wooteco.subway.admin.service.LineStationService;
 
 @RestController
@@ -49,5 +53,17 @@ public class LineStationController {
         @PathVariable Long stationId) {
         lineStationService.removeStationFromLine(lineId, stationId);
         return ResponseEntity.noContent().build();
+    }
+
+    @ExceptionHandler(NotFoundException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ApiError notFoundException(NotFoundException exception) {
+        return new ApiError(HttpStatus.BAD_REQUEST, exception.getMessage());
+    }
+
+    @ExceptionHandler(LineStationException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ApiError lineStationException(LineStationException exception) {
+        return new ApiError(HttpStatus.BAD_REQUEST, exception.getMessage());
     }
 }
