--- src/main/java/wooteco/subway/admin/controller/advice/ControllerAdvice.java
@@ -1,31 +1,0 @@
-package wooteco.subway.admin.controller.advice;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import wooteco.subway.admin.exception.DuplicateNameException;
-import wooteco.subway.admin.exception.LineStationException;
-import wooteco.subway.admin.exception.NotFoundException;
-
-@RestControllerAdvice
-public class ControllerAdvice {
-    @ExceptionHandler(DuplicateNameException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ApiError existingNameException(DuplicateNameException exception) {
-        return new ApiError(HttpStatus.BAD_REQUEST, exception.getMessage());
-    }
-
-    @ExceptionHandler(NotFoundException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ApiError notFoundException(NotFoundException exception) {
-        return new ApiError(HttpStatus.BAD_REQUEST, exception.getMessage());
-    }
-
-    @ExceptionHandler(LineStationException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ApiError lineStationException(LineStationException exception) {
-        return new ApiError(HttpStatus.BAD_REQUEST, exception.getMessage());
-    }
-}
