+++ src/main/java/wooteco/subway/admin/dto/request/LineRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.dto;
+package wooteco.subway.admin.dto.request;
 
 import java.time.LocalTime;
