+++ src/main/java/wooteco/subway/admin/dto/request/StationCreateRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.dto;
+package wooteco.subway.admin.dto.request;
 
 import wooteco.subway.admin.domain.Station;
