+++ src/main/java/wooteco/subway/admin/dto/resopnse/LineDetailResponse.java
@@ -1,13 +1,13 @@
-package wooteco.subway.admin.dto;
+package wooteco.subway.admin.dto.resopnse;
 
 import java.time.LocalDateTime;
 import java.time.LocalTime;
-import java.util.Objects;
+import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import wooteco.subway.admin.domain.Line;
 
-public class LineResponse {
+public class LineDetailResponse {
     private Long id;
     private String name;
     private String color;
@@ -19,11 +19,15 @@ public class LineResponse {
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 
-    private LineResponse() {
+    private List<StationResponse> stations;
+
+    private LineDetailResponse() {
     }
 
-    public LineResponse(Long id, String name, String color, LocalTime startTime,
-        LocalTime endTime, int intervalTime, LocalDateTime createdAt, LocalDateTime updatedAt) {
+    public LineDetailResponse(Long id, String name, String color, LocalTime startTime,
+        LocalTime endTime,
+        int intervalTime,
+        LocalDateTime createdAt, LocalDateTime updatedAt, List<StationResponse> stations) {
         this.id = id;
         this.name = name;
         this.color = color;
@@ -32,11 +36,14 @@ public class LineResponse {
         this.intervalTime = intervalTime;
         this.createdAt = createdAt;
         this.updatedAt = updatedAt;
+        this.stations = stations;
     }
 
-    public static LineResponse from(Line line) {
-        return new LineResponse(line.getId(), line.getName(), line.getColor(), line.getStartTime(),
-            line.getEndTime(), line.getIntervalTime(), line.getCreatedAt(), line.getUpdatedAt());
+    public static LineDetailResponse of(Line line, List<StationResponse> stations) {
+        return new LineDetailResponse(line.getId(), line.getName(), line.getColor(),
+            line.getStartTime(),
+            line.getEndTime(),
+            line.getIntervalTime(), line.getCreatedAt(), line.getUpdatedAt(), stations);
     }
 
     public Long getId() {
@@ -63,31 +70,15 @@ public class LineResponse {
         return intervalTime;
     }
 
+    public List<StationResponse> getStations() {
+        return stations;
+    }
+
     public LocalDateTime getCreatedAt() {
         return createdAt;
     }
 
     public LocalDateTime getUpdatedAt() {
         return updatedAt;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o)
-            return true;
-        if (o == null || getClass() != o.getClass())
-            return false;
-        LineResponse that = (LineResponse)o;
-        return intervalTime == that.intervalTime &&
-            Objects.equals(id, that.id) &&
-            Objects.equals(name, that.name) &&
-            Objects.equals(color, that.color) &&
-            Objects.equals(startTime, that.startTime) &&
-            Objects.equals(endTime, that.endTime);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, name, color, startTime, endTime, intervalTime);
     }
 }
