+++ src/main/java/wooteco/subway/admin/dto/resopnse/StationResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.dto;
+package wooteco.subway.admin.dto.resopnse;
 
 import java.time.LocalDateTime;
 import java.util.Objects;
