+++ src/main/java/wooteco/subway/admin/service/LineService.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import wooteco.subway.admin.domain.Line;
-import wooteco.subway.admin.dto.LineResponse;
+import wooteco.subway.admin.dto.resopnse.LineResponse;
 import wooteco.subway.admin.exception.DuplicateNameException;
 import wooteco.subway.admin.exception.NotFoundException;
 import wooteco.subway.admin.repository.LineRepository;
