+++ src/test/java/wooteco/subway/admin/acceptance/StationAcceptanceTest.java
@@ -17,7 +17,7 @@ import org.springframework.test.context.jdbc.Sql;
 
 import io.restassured.RestAssured;
 import io.restassured.specification.RequestSpecification;
-import wooteco.subway.admin.dto.StationResponse;
+import wooteco.subway.admin.dto.resopnse.StationResponse;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Sql("/truncate.sql")
