+++ src/test/java/wooteco/subway/admin/service/LineServiceTest.java
@@ -16,7 +16,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import wooteco.subway.admin.domain.Line;
-import wooteco.subway.admin.dto.LineResponse;
+import wooteco.subway.admin.dto.resopnse.LineResponse;
 import wooteco.subway.admin.exception.DuplicateNameException;
 import wooteco.subway.admin.exception.NotFoundException;
 import wooteco.subway.admin.repository.LineRepository;
